+++ SOURCE_CODE/Properties/AssemblyInfo.cs
@@ -7,8 +7,8 @@ using System.Runtime.InteropServices;
 [assembly: AssemblyProduct("ServoERP")]
 [assembly: AssemblyCopyright("Copyright (c) Harshal Sonawane 2026")]
 [assembly: ComVisible(false)]
-[assembly: AssemblyVersion("1.1.335.0")]
-[assembly: AssemblyFileVersion("1.1.335.0")]
+[assembly: AssemblyVersion("1.1.336.0")]
+[assembly: AssemblyFileVersion("1.1.336.0")]